Home screen added.
