+++ ArabamAssignment/App/Delegate/AppDelegate.swift
@@ -1,0 +1,31 @@
+//
+//  AppDelegate.swift
+//  ArabamAssignment
+//
+//  Created by Furkan Arslan on 25.01.2021.
+//
+
+import UIKit
+
+@main
+class AppDelegate: UIResponder, UIApplicationDelegate {
+
+    var window: UIWindow?
+
+    func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
+        
+        setupWindow()
+        
+        return true
+    }
+    
+    private func setupWindow() {
+        let window = UIWindow(frame: UIScreen.main.bounds)
+        let viewController = HomeViewController()
+        let navigationController = UINavigationController(rootViewController: viewController)
+        window.rootViewController = navigationController
+        window.makeKeyAndVisible()
+        self.window = window
+    }
+}
+
