--- ArabamAssignment/AppDelegate.swift
@@ -1,21 +1,0 @@
-//
-//  AppDelegate.swift
-//  ArabamAssignment
-//
-//  Created by Furkan Arslan on 25.01.2021.
-//
-
-import UIKit
-
-@main
-class AppDelegate: UIResponder, UIApplicationDelegate {
-
-    var window: UIWindow?
-
-    func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
-        // Override point for customization after application launch.
-        return true
-    }
-
-}
-
