+++ ArabamAssignment/Screens/Home/Controller/HomeViewController.swift
@@ -1,0 +1,17 @@
+//
+//  HomeViewController.swift
+//  ArabamAssignment
+//
+//  Created by Muhammed Karakul on 9.01.2022.
+//
+
+import UIKit
+
+final class HomeViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        title = "Home"
+    }
+}
