--- ArabamAssignment/ViewController.swift
@@ -1,19 +1,0 @@
-//
-//  ViewController.swift
-//  ArabamAssignment
-//
-//  Created by Furkan Arslan on 25.01.2021.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view.
-    }
-
-
-}
-
